back-end/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
facenet-pytorch==2.6.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
supabase==2.0.2
python-dotenv==1.0.0 